version 0.3.0 * Finally and totally obliterated the src-expresskeysconf directory. Such a tool will never be written by me. But there is hope balancing on the horizon! At http://alavaliant.googlepages.com you'll find "wacom-config", which is a pygtk (2.8+) point-and-click program for configuration of the linuxwacom driver (through "xsetwacom") and expresskeys version 0.2.x config file editing.
No more manual wrangling with "xprop" and "xev"! A future version
will support expresskeys 0.3.x (I've been assured).

* New configuration file (version 3), incompatible with version 2. The
reasoning around this switch can be read about in the USAGE file, ca
2/3 down under the heading "NEW in expresskeys version 0.3.0".

This will hopefully be the _last_ config file change that impacts the
user in a negative way. Future changes can be done non-destructively.

* New configuration _file name_. Depending on devices identified, either
a "padless.conf1" or an "intuos3.conf1" file is written out. Apart from
only containing relevant record fields, the headers of these files are
tagged with device "Identifier" strings from xorg.conf, like:

_intuos3.conf1_                         _padless.conf1_

ConfigVersion 3                         ConfigVersion 3
Identifier1Pad 1stIntuos3Pad            Identifier1Pad LACKING
Identifier1Sty 1stIntuos3Stylus1        Identifier1Sty Stylus1

The TODO file - better known as the "Wish-list" - hints towards why this
filename change has become desirable...

* Automatic detection and use of the first 'pad' and/or 'stylus' device found
through the XListInputDevices call. This happens unconditionally, but can be
overridden by specifically naming a device on the command line.

We still only allow one device of each kind (for now) so eg an automatically
picked stylus, say 1stIntuos3Stylus1, gets dropped if the user specifies the
1stIntuos3Stylus2 or a 2ndIntuos3Stylus1.

Anyone with a little programming understanding can change a copy of the code
in globals.c and alter the config_dir string to eg "/.expresskeys2". Then
that program copy can support another set of pad/stylus (until expresskeys
properly handles multi-devices).

* Expresskeys can now trace the stylus usage and automatically change
the pressure sensitivity (PressCurve) depending on which program window
that has focus. We call "xsetwacom" for the actual change, but the
implementation has minimal overhead - can't be felt or seen on my machine:

NOTE: To partially overcome a limitation in any Qt3 program (see the BUGS
file; they swallow all stylus button presses) we also have to judge the
state _every time_ the pen comes within operational height of the tablet,
aka ProximityIn.

1a) Stylusbutton - tip or rocker - pressed (doesn't matter which),
1b) or stylus reports itself as having entered proximity of the tablet.
2) Is the current focuswindow in our program list (plus "default")? = True.
3) Does the program name differ from the name in a historybuffer (1 back)?
4) Does the program presscurve differ from a historybuffer curve (1 back)?
5) Call xsetwacom only if these are true.

The above mentioned USAGE entry has all the important details.

So... now the program enters a more generic stage where it's beneficial
to any Wacom tablet user. Though, the name will remain expresskeys ;-)

* Deleted the old-extra directory with the small scripts for re-reading the
configuration file or terminating the program.

It is actually quite embarrassing, but the two options have been available
from _within_ the main_setup.c file since... forever. We already did a signal
send to check for a running daemon, so all it took was a slight rearrangement
and a few extra code lines. I _am_ myopic, but this oversight borders on
blindness.

It is now possible to do an "expresskeys -r" or "expresskeys -k" against
a daemon instance. (BLUSH)!

* A few more command line switches have been introduced:

-x sets -v and exits after some important info blocks.
-s tells a daemon instance to report status (info block).

The 'status' signal is triggered through SIGUSR2 (re-read of the config file
has always been linked with SIGUSR1). The information printed out by -s
is almost exactly how -x displays it during a non-daemon run. Difference
being that -s also gives the "OUR RUN-PID" field.

OBSERVE: If expresskeys is started from the "outside" by use of .xinitrc or
similar methods, the output of -s will be seen on that first text terminal
(most often back at "Ctrl-Alt-F1" when in X). To get visibility from a terminal
_inside_ X, first do an "expresskeys -k" then "expresskeys -d" and THEN the
"expresskeys -s" ;-)

* Finally remembered to give a warning about Gimp and the pad device.
The USAGE file has it as a top item, and I'll copy the text here:

"Important: Gimp doesn't know _anything_ about a "pad" device, so the option
File --> Preferences --> Input Devices --> Configure Extended Input Devices...
--> [Device: pad  Mode: Disabled] is how it should be set. If you change it
to Sceen or Window, pressing pad buttons or using touch strips will NOT work."

* Updated to a working email address in the AUTHORS file and removed the
obsolete private project page from README. The canonical page now is:

http://freshmeat.net/projects/wacomexpresskeys

It has been so since version 0.2.1, I just never got around to mentioning it.

+++ src-expresskeys/on_signal.c
@@ -1,18 +1,18 @@
 /*
  on_signal.c -- Support ExpressKeys & Touch Strips on a Wacom Intuos3 tablet.
- 
+
  Copyright (C) 2005-2006 - Mats Johannesson
- 
+
  This program is free software; you can redistribute it and/or modify
  it under the terms of the GNU General Public License as published by
  the Free Software Foundation; either version 2 of the License, or
  (at your option) any later version.
- 
+
  This program is distributed in the hope that it will be useful,
  but WITHOUT ANY WARRANTY; without even the implied warranty of
  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  GNU General Public License for more details.
- 
+
  You should have received a copy of the GNU General Public License
  along with this program; if not, write to the Free Software
  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307, USA.
@@ -21,20 +21,21 @@
 #include "globals.h"
 
 /*+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- Function acts as a signal handler replacement for SIGUSR1 and SIGUSR2
- On these signals we read a possibly changed configuration file again
+ Function acts as a signal handler replacement for SIGUSR1
+ On this signal we read a possibly changed configuration file again
  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++*/
 
 void re_read_file_config(int signum)
 {
 
 	struct program *p;
+	struct configstrings *hp;
 
 	FILE *fp;
 	FILE *errorfp;
 
 /* Open (and truncate) an error log for future reference */
-	
+
 	if ((errorfp = fopen(total_error_file, "w")) == NULL) {
 		exit_on_error(errorfp, "%s ERROR: Reread - Can not open %s in write mode\n", our_prog_name, total_error_file);
 	}
@@ -42,34 +43,80 @@ void re_read_file_config(int signum)
 /* Read in an existing configuration file */
 
 	p = external_list;
+	hp = human_readable;
 	if ((fp = fopen(total_config_file, "r")) == NULL) {
 		exit_on_error(errorfp, "%s ERROR: Reread - Can not open %s in read mode\n", our_prog_name, total_config_file);
 	} else {
-		switch (read_file_config((void *)&p, fp)){
-			
-			case 0:
+		switch (read_file_config((void *)&p, (void *)&hp, fp)){
+
+			case 0: /* No errors */
 			fclose(fp);
-			break;
+			break; /* OBS An identical list of error code returns exist in the
+					main_setup.c file. So change both when altering. */
 
 			case 1:
 			fclose(fp);
 			exit_on_error(errorfp, "%s ERROR: Reread - No complete record found in %s\n", our_prog_name, total_config_file);
-			
+
 			case 2:
 			fclose(fp);
 			exit_on_error(errorfp, "%s ERROR: Reread - Memory allocation error while parsing %s\n", our_prog_name, total_config_file);
 
 			case 3:
 			fclose(fp);
-			exit_on_error(errorfp, "%s ERROR: Reread - Config File Version %d not found\n", our_prog_name, (void *)CONFIG_VERSION);
+			exit_on_error(errorfp, "%s ERROR: Reread - Config File Version 3 or higher not found\n", our_prog_name, "");
+
+			case 4:
+			fclose(fp);
+			exit_on_error(errorfp, "%s ERROR: Reread - A line was too long to handle in the Config File %s\n", our_prog_name, total_config_file);
+
+			case 5:
+			fclose(fp);
+			exit_on_error(errorfp, "%s ERROR: Reread - A program record named \"default\" must exist in %s\n", our_prog_name, total_config_file);
 
 			default:
 			fclose(fp);
 			exit_on_error(errorfp, "%s ERROR: Reread - Unknown error while parsing %s\n", our_prog_name, total_config_file);
 		}
 	}
-	
+
 	fclose(errorfp);
+}
+
+/*+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+ Function acts as a signal handler replacement for SIGUSR2
+ On this signal we print some choise information to the screen
+ +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++*/
+
+void status_report(int signum)
+{
+
+	struct program *p;
+	p = external_list;
+	int i;
+
+	fprintf(stderr, "PGR VERSION = %s\n", our_prog_version);
+	fprintf(stderr, "USR HOMEDIR = %s\n", getenv("HOME"));
+	fprintf(stderr, "OUR CNF-DIR = %s\n", total_config_dir);
+	fprintf(stderr, "OUR CNFFILE = %s\n", total_config_file);
+	fprintf(stderr, "OUR PIDFILE = %s\n", total_pid_file);
+	fprintf(stderr, "OUR LOGFILE = %s\n", total_error_file);
+	if (pad1_name) {
+		fprintf(stderr, "OUR PD1NAME = %s\n", pad1_name);
+	}
+	if (stylus1_name) {
+		fprintf(stderr, "OUR ST1NAME = %s\n", stylus1_name);
+	}
+	fprintf(stderr, "%s-user = %d\n", configstring, userconfigversion);
+	fprintf(stderr, "%s-ours = %s\n", configstring, configversion);
+	fprintf(stderr, "ConfigHeaderFields = %d\n", configheaderfields);
+	for (i = 0; i < num_list; i++, p++) {
+		fprintf(stderr, "PGR RECNAME = %s\n", p->class_name);
+		fprintf(stderr, "ST1 PRCURVE = \"%s\"\n", p->stylus1_presscurve);
+	}
+	fprintf(stderr, "PGR RECORDS = %d\n", num_list);
+	fprintf(stderr, "OUR RUN-PID = %d\n", getpid());
+
 }
 
 /*+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
@@ -92,8 +139,8 @@ void clean_up_exit(int signum)
 		raise(signum);
 	}
 	clean_up_exit_in_progress = 1;
-	
-	if (go_daemon) {
+
+	if ((go_daemon) && (!second_instance)) { /* Prevent accidental deletion */
 		unlink(total_pid_file);
 	}
 
@@ -115,21 +162,23 @@ void clean_up_exit(int signum)
 
 	if (num_list) {
 		for (i = 0; i < num_list; i++, p++) {
+			free(p->stylus1_presscurve);
 			free(p->class_name);
 		}
 	}
 
-	if (pad_info_block) {
-		XFreeDeviceList(pad_info_block);
+	if (pad1_info_block) {
+		XFreeDeviceList(pad1_info_block);
 	}
 
-	if (pen_info_block) {
-		XFreeDeviceList(pen_info_block);
+	if (stylus1_info_block) {
+		XFreeDeviceList(stylus1_info_block);
 	}
 
-/* pad_device and pen_device should not be explicitly closed by a
+/* pad1_device and stylus1_device should not be explicitly closed by a
    call to XCloseDevice. It leaves a message from X (in the terminal
-   where X is started from) saying: "ProcXCloseDevice to close or not ?" */
+   where X is started from) saying: "ProcXCloseDevice to close or not ?"
+   Like the program "xsetwacom" does on every device change...*/
 
 	if (display) {
 		XCloseDisplay(display);
